# docs/swagger.yaml
definitions:
...
  models.UserResponse:
    properties:
      createdAt:
        type: integer
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      isAdmin:
        type: boolean
      lastName:
        type: string
      picture:
        type: string
      updatedAt:
        type: integer
      verifiedEmail:
        type: boolean
    required:
    - firstName
    - lastName
    type: object
  swagger.AccessForbiddenResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: Sorry you don't have access to visit this page!
        type: string
      ok:
        example: false
        type: boolean
    type: object
...
  swagger.EmptyData:
    type: object
...
  swagger.NotFoundResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: 404 not found!
        type: string
      ok:
        example: false
        type: boolean
    type: object
  swagger.PaginateUsersData:
    properties:
      count:
        example: 10
        type: integer
      list:
        items:
          $ref: '#/definitions/models.UserResponse'
        type: array
    type: object
...
  swagger.SingleUserResponse:
    properties:
      data:
        $ref: '#/definitions/models.UserResponse'
      msg:
        example: Successful message
        type: string
      ok:
        example: true
        type: boolean
    type: object
...
  swagger.UnauthenticatedResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: You must login first!
        type: string
      ok:
        example: false
        type: boolean
    type: object
  swagger.UsersListResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.PaginateUsersData'
      msg:
        example: Successful message
        type: string
      ok:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /admin/users:
    get:
      consumes:
      - application/json
      description: list of paginated response , authentication required
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.UsersListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthenticatedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.AccessForbiddenResponse'
      summary: get users list
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: create user and admin , admin only
      parameters:
      - description: unique email
        in: formData
        name: email
        required: true
        type: string
      - description: 'password that have at least 8 length and contain an alphabet
          and number '
        in: formData
        name: password
        required: true
        type: string
      - description: 'repeatPassword that have at least 8 length and contain an alphabet
          and number '
        in: formData
        name: repeatPassword
        required: true
        type: string
      - description: firstName
        in: formData
        name: firstName
        required: true
        type: string
      - description: lastName
        in: formData
        name: lastName
        required: true
        type: string
      - description: isAdmin
        in: formData
        name: isAdmin
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.UsersListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthenticatedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.AccessForbiddenResponse'
      summary: create users
      tags:
      - admin
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: delete user or admin , admin only
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.UsersListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthenticatedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.AccessForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.NotFoundResponse'
      summary: delete user
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: update user and admin , admin only
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: unique email
        in: formData
        name: email
        required: true
        type: string
      - description: firstName
        in: formData
        name: firstName
        required: true
        type: string
      - description: lastName
        in: formData
        name: lastName
        required: true
        type: string
      - description: isAdmin
        in: formData
        name: isAdmin
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SingleUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthenticatedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.AccessForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.NotFoundResponse'
      summary: update user
      tags:
      - admin
...
swagger: "2.0"